Embedding-ZSL-baseline/requirements.txt
pandas>=2.2
openpyxl>=3.1
pyyaml>=6.0
tqdm>=4.66
scikit-learn>=1.5
sentence-transformers>=3.0
numpy>=1.26
rich>=13.7
requests>=2.32
